Added script to create files, modified previous filenames

+++ H1/datadb100_h1_t10_150974219.sql
@@ -7,7 +7,7 @@ CREATE TABLE shoe (
     id int NOT NULL,
     name VARCHAR NOT NULL,
     type VARCHAR,
-    price int,
+    price DECIMAL,
     PRIMARY KEY (id),
     UNIQUE (name)
 );
